chore, fix, refactor: ruff linting checks, minimal fixes and file movements
