+++ src/dotcop/data/interface/ConfigFileInterface.py
@@ -13,13 +13,13 @@ class ConfigFileInterface(ABC):
         """ Retrieve the 'log_path' field from the configuration file"""
 
     @abstractmethod
-    def get_package_path(self) -> str: 
+    def get_package_path(self) -> str:
         """ Retrieve the 'package_path' field from the configuration file"""
-    
+
     @abstractmethod
-    def get_database_path(self) -> str: 
+    def get_database_path(self) -> str:
         """ Retrieve the 'database_path' field from the configuration file"""
 
     @abstractmethod
-    def get_meta_path(self) -> str: 
+    def get_meta_path(self) -> str:
         """ Retrieve the 'meta_path' field from the configuration file"""
